Split GUI status bar tests
